requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
DateTime==5.5
distlib==0.3.8
docker==7.1.0
dos2unix==1
dulwich==0.21.7
environs==10.0.0
fastjsonschema==2.19.1
filelock==3.13.4
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
h11==0.14.0
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
lxml==5.2.1
magic-filter==1.0.12
marshmallow==3.20.1
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.1
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==8.0.0
python-dotenv==1.0.1
pytz==2024.1
rapidfuzz==3.8.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
schedule==1.2.2
selenium==4.21.0
shellingham==1.5.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
StrEnum==0.4.15
tomlkit==0.12.4
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2024.4.10
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.26.0
wsproto==1.2.0
xattr==1.1.0
yarl==1.9.4
zipp==3.18.1
zope.interface==6.4.post2
webdriver_manager==4.0.1
greenlet==3.0.3
playwright==1.44.0
pyee==11.1.0